+++ geoandino/apps/datajsonar/managers.py
