+++ geoandino/apps/datajsonar/urls.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+from django.conf.urls import patterns, url
+
+from .views import datajsonar
+
+urlpatterns = patterns('',
+   url(r'^data.json$', datajsonar, name="datajsonar"),
+ )
