+++ geoandino/apps/datajsonar/views.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from django.http import JsonResponse
-from geoandino.utils.datajsonar import data_jsonar
+from .utils.datajsonar import data_jsonar
 
-def data_json(request):
+def datajsonar(request):
     return JsonResponse(data_jsonar())
